modifications to make game logging more sensible. all logging now happens through one of the many log_...() functions in log.c. if you need to log something, but it doesn't fit into a category, just use log_message(). It takes variable arguments, so you can:
  log_message("just a string");
or
  log_message("%d is the id of the player", player);

This code has been tested with full command and debug logging.
It doesn't appear to crash the game, and there is no more odd
output from Seeth when you gripe!

this is a pre-cursor to the New, Funky Log-Handling Perl Modules.
watch this space. or one quite near to it.